Fetch build statuses
Big ol' refactor to fetch packages which have failed to build in a
packageset

To find out which, look for the log "build not successful".

It was interminably slow otherwise, so this refactor adds a thread pool
to access LP concurrently. There will be per-thread caches now. Probably
do expect some bugs here.