Move thread-local LP instances to consumers
There was a crazy memory leak that never seemed to end. Instead of
trying to find the top of it, this fixes it. Not exactly sure why but we
were churning Launchpad instances like mad and they weren't being
collected by the GC. Moving the handling of this out of the Launchpad
class to the consumers of that class fixes that, memory usage is stable
and low now.